+++ RELEASES.md
@@ -1,5 +1,6 @@
 Version 0.8.12 (202X-XX-XX)
 ========================
+ * Volkswagen T-Roc 2021 support thanks to jyoung8607!
 
 Version 0.8.11 (2021-11-22)
 ========================